Maha_OOD/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.9.0=pypi_0
aiohttp=3.6.2=pypi_0
alembic=1.3.1=pypi_0
altair=4.0.1=py_0
annoy=1.16.3=pypi_0
astor=0.8.1=pypi_0
async-timeout=3.0.1=pypi_0
attrs=19.3.0=py_0
backcall=0.1.0=py37_0
bcolz=1.2.1=py37h04863e7_0
beautifulsoup4=4.9.0=pypi_0
blas=1.0=mkl
bleach=3.1.0=py37_0
blessings=1.7=pypi_0
blosc=1.16.3=hd408876_0
bokeh=2.0.1=py37_0
boto3=1.12.38=pypi_0
botocore=1.15.38=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2020.1.1=0
cachetools=4.0.0=pypi_0
certifi=2020.4.5.1=py37_0
cffi=1.13.2=py37h2e261b9_0
chardet=3.0.4=pypi_0
click=7.1.1=py_0
cliff=3.1.0=pypi_0
cloudpickle=1.3.0=py_0
cmaes=0.4.0=pypi_0
cmd2=0.8.9=pypi_0
colorama=0.4.3=pypi_0
colorlog=4.1.0=pypi_0
configparser=5.0.0=pypi_0
cudatoolkit=10.1.243=h6bb024c_0
cyrus-sasl=2.1.26=h82bf5a1_4
cytoolz=0.10.1=py37h7b6447c_0
dask=2.14.0=py_0
dask-core=2.14.0=py_0
databricks-cli=0.10.0=pypi_0
dbus=1.13.12=h746ee38_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
distributed=2.14.0=py37_0
docker=4.2.0=pypi_0
docopt=0.6.2=pypi_0
docutils=0.15.2=pypi_0
entrypoints=0.3=py37_0
expat=2.2.6=he6710b0_0
filelock=3.0.12=pypi_0
flask=1.1.2=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fsspec=0.7.1=py_0
future=0.18.2=pypi_0
git=2.23.0=pl526hacde149_0
gitdb=4.0.2=pypi_0
gitpython=3.1.0=pypi_0
glib=2.63.1=h5a9c865_0
gmp=6.1.2=h6c8ec71_1
google=2.0.3=pypi_0
google-auth=1.13.1=pypi_0
google-auth-oauthlib=0.4.1=pypi_0
gorilla=0.3.0=pypi_0
gpustat=0.6.0=pypi_0
grpcio=1.28.1=pypi_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
gunicorn=20.0.4=pypi_0
hdf5=1.10.4=hb1b8bf9_0
heapdict=1.0.1=py_0
hydra-core=0.11.3=pypi_0
icu=58.2=h9c2bf20_1
idna=2.9=pypi_0
importlib_metadata=0.23=py37_0
intel-openmp=2019.4=243
ipykernel=5.1.3=py37h39e3cac_0
ipython=7.9.0=py37h39e3cac_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
itsdangerous=1.1.0=pypi_0
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
jmespath=0.9.5=pypi_0
joblib=0.14.1=py_0
jpeg=9b=h024ee3a_2
jsonnet=0.11.2=py37he1b5a44_0
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.0.0=py37_0
jupyter_core=4.6.1=py37_0
krb5=1.16.4=h173b8e3_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libcurl=7.67.0=h20c2e04_0
libdb=6.1.26=he6710b0_0
libedit=3.1.20181209=hc058e9b_0
libevent=2.1.10=h72c5cf5_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
liblapack=3.8.0=14_mkl
libntlm=1.5=h7b6447c_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.16=h1bed415_0
libssh2=1.8.2=h1ba5d50_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
locket=0.2.0=py37_1
lz4-c=1.8.1.2=h14c3975_0
lzo=2.10=h49e0be7_2
mako=1.1.0=pypi_0
markdown=3.2.1=pypi_0
markupsafe=1.1.1=py37h7b6447c_0
mistune=0.8.4=py37h7b6447c_0
mkl=2019.4=243
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
mlflow=1.7.2=pypi_0
mock=4.0.1=py_0
more-itertools=7.2.0=py37_0
msgpack-python=1.0.0=py37hfd86e86_1
multidict=4.7.5=pypi_0
mysql=5.7.24=hbb652a2_0
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py37_0
ncurses=6.1=he6710b0_1
ninja=1.9.0=py37hfd86e86_0
nltk=3.4.5=py37_0
nose=1.3.7=pypi_0
notebook=6.0.2=py37_0
numexpr=2.7.1=py37h423224d_0
numpy=1.18.1=py37h8960a57_1
nvidia-ml-py3=7.352.0=pypi_0
oauthlib=3.1.0=pypi_0
olefile=0.46=py37_0
omegaconf=1.4.1=pypi_0
openssl=1.1.1f=h7b6447c_0
optuna=1.3.0=pypi_0
packaging=20.3=py_0
pandas=1.0.3=py37h0da4684_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.5.1=py_0
partd=1.1.0=py_0
pbr=5.4.5=pypi_0
pcre=8.43=he6710b0_0
perl=5.26.2=h14c3975_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.1.1=pypi_0
pip=19.3.1=py37_0
prettytable=0.7.2=pypi_0
prometheus-flask-exporter=0.13.0=pypi_0
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
protobuf=3.11.3=pypi_0
psutil=5.7.0=py37h7b6447c_0
ptyprocess=0.6.0=py37_0
py-spy=0.3.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.19=py37_0
pydevd-pycharm=201.6668.115=pypi_0
pygments=2.4.2=py_0
pyparsing=2.4.6=py_0
pyperclip=1.8.0=pypi_0
pyqt=5.9.2=py37h05f1152_2
pyrsistent=0.15.6=py37h7b6447c_0
pytables=3.6.1=py37h71ec239_0
python=3.7.5=h0371630_0
python-dateutil=2.8.1=py_0
python-editor=1.0.4=pypi_0
python_abi=3.7=1_cp37m
pytorch-lightning=0.7.3=pypi_0
pytorch-transformers=1.2.0=pypi_0
pytz=2019.3=py_0
pyyaml=5.3.1=pypi_0
pyzmq=18.1.0=py37he6710b0_0
qt=5.9.7=h5867ecd_1
qtconsole=4.6.0=py_0
querystring-parser=1.2.4=pypi_0
ray=0.8.4=pypi_0
readline=7.0=h7b6447c_5
redis=3.4.1=pypi_0
regex=2020.4.4=pypi_0
requests=2.23.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.0=pypi_0
s3transfer=0.3.3=pypi_0
sacremoses=0.0.38=pypi_0
scikit-learn=0.22.2.post1=pypi_0
scipy=1.4.1=py37h0b6359f_0
send2trash=1.5.0=py37_0
sentencepiece=0.1.85=pypi_0
setuptools=42.0.1=py37_0
simplejson=3.17.0=pypi_0
sip=4.19.8=py37hf484d3e_0
six=1.13.0=py37_0
smmap=3.0.1=pypi_0
snappy=1.1.7=hbae5bb6_3
sortedcontainers=2.1.0=py37_0
soupsieve=2.0=pypi_0
sqlalchemy=1.3.11=pypi_0
sqlite=3.30.1=h7b6447c_0
sqlparse=0.3.1=pypi_0
stevedore=1.32.0=pypi_0
sweeper=0.9.0=pypi_0
tabulate=0.8.7=pypi_0
tblib=1.6.0=py_0
tensorboard=2.2.0=pypi_0
tensorboard-plugin-wit=1.6.0.post3=pypi_0
tensorly=0.4.5=pypi_0
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=hbc83047_0
tmux=2.9=ha1ba12b_1
tokenizers=0.5.2=pypi_0
toolz=0.10.0=py_0
torch=1.3.1=pypi_0
torchvision=0.4.2=pypi_0
tornado=6.0.3=py37h7b6447c_0
tqdm=4.45.0=pypi_0
traitlets=4.3.3=py37_0
transformers=2.8.0=pypi_0
typing_extensions=3.7.4.1=py37_0
urllib3=1.25.8=pypi_0
vega_datasets=0.8.0=py_0
wcwidth=0.1.7=py37_0
webencodings=0.5.1=py37_1
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.33.6=py37_0
widgetsnbextension=3.5.1=py37_0
xz=5.2.4=h14c3975_4
yaml=0.1.7=had09818_2
yarl=1.4.2=pypi_0
zeromq=4.3.1=he6710b0_3
zict=2.0.0=py_0
zipp=0.6.0=py_0
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
